split dygraph and static
