+++ python/paddle/incubate/optimizer/functional/line_search.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .utils import _value_and_gradient
+from utils import _value_and_gradient
 import paddle
